moved player to its own module
